+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- Text alignment
+- Text alignment and indent 
 
 ## [0.2.0] - 2025-03-17
 ### Added